astronomer/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos[dbt-bigquery]==1.0.3
protobuf==3.20.0
scikit-learn
pandas
google-cloud-bigquery
numpy